NU-2405 move table actions to custom hook
